requirements.txt
fastmcp
openai
openai-agents
python-dotenv